% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTRF.R
\docType{data}
\name{HTRF}
\alias{HTRF}
\title{Human time-restricted feeding data}
\format{
A data frame with 13167 rows and 125 variables. Each row is gene expression value. Each column represents one sample.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE129843}
}
\usage{
HTRF
}
\description{
Short-Term Time-Restricted Feeding Alters Rhythmicity of Lipid and Amino Acid Metabolites in Overweight Men (Human time-restricted feeding data)
}
\references{
Lundell LS, Parr EB, Devlin BL, Ingerslev LR et al. Time-restricted feeding alters lipid and amino acid metabolite rhythmicity without perturbing clock gene expression. Nat Commun 2020 Sep 16;11(1):4643. PMID: 32938935
}
\keyword{datasets}
