% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Subjects information for Human time-restricted feeding data}
\format{
A data frame with 69 rows and 4 columns. Each row represents one subject. First column is subject ID. Second row is phase. Third row is treatment group. Fourth row is time.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE129843}
}
\usage{
pheno
}
\description{
Subjects information for Short-Term Time-Restricted Feeding Alters Rhythmicity of Lipid and Amino Acid Metabolites in Overweight Men (Human time-restricted feeding data)
}
\references{
Lundell LS, Parr EB, Devlin BL, Ingerslev LR et al. Time-restricted feeding alters lipid and amino acid metabolite rhythmicity without perturbing clock gene expression. Nat Commun 2020 Sep 16;11(1):4643. PMID: 32938935.
}
\keyword{datasets}
