% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_diff.R
\name{rpt_diff}
\alias{rpt_diff}
\title{Likelihood-based Tests for Detecting Differential Circadian Pattern with Repeated Measurement}
\usage{
rpt_diff(tt1, yy1, tt2, yy2, id1, id2, group, period = 24, method = "LR")
}
\arguments{
\item{tt1}{Time vector of condition 1.}

\item{yy1}{Expression vector of condition 1.}

\item{tt2}{Time vector of condition 2.}

\item{yy2}{Expression vector of condition 2.}

\item{period}{Period of the since curve. Default is 24.}

\item{method}{Test used to detect differential circadian pattern. It can be chosen either "LR" or "F". Default is LR.}
}
\value{
A list, see details below.

Formula 1: \eqn{yy = amp * sin(2\pi/period * (phase + tt)) + offset.}
Formula 2: \eqn{yy = A * sin(2\pi/period * tt) + B * cos(2*\pi/period * tt) + offset.}
\item{stat}{Test statistic.}
\item{pvalue}{P-value from the test.}
}
\description{
Likelihood-based test for differential circadian pattern detection with repeated measurement.
}
\details{
Test differential rhythmicity of circadian curve fitting using random intercept linear mixed model and likelihood-based tests.
}
\examples{
Example 1

set.seed(32611)
n1 <- 12
n2 <- 12
m1 <- 10
m2 <- 10
id1 <- rep(1:n1,each=m1)
id2 <- rep((1+n1):(n1+n2),each=m2)
rho <- 0.2
offset1 <- runif(1,0,3)
offset2 <- runif(1,2,5)
amp1 <- 1
amp2 <- 2
phase1 <- 3
phase2 <- 6
sigmaMat1 <- ifelse(diag(m1)==1,1,rho)
sigmaMat2 <- ifelse(diag(m2)==1,1,rho)
t1 <- runif(m1,0,24)
tt1 <- rep(t1,n1)
yy1 <- as.vector(t(mvrnorm(n1,amp1*sin(2*pi/24*(phase1+t1))+offset1,sigmaMat1)))
t2 <- runif(m2,0,24)
tt2 <- rep(t2,n2)
yy2 <- as.vector(t(mvrnorm(n2,amp2*sin(2*pi/24*(phase2+t2))+offset2,sigmaMat2)))
group <- c(rep("group1",n1*m1),rep("group2",n2*m2))
rpt_diff(tt1,yy1,tt2,yy2,id1,id2,group)


Example 2

set.seed(32611)
n1 <- 12
n2 <- 12
m1 <- 10
m2 <- 10
id1 <- rep(1:n1,each=m1)
id2 <- rep((1+n1):(n1+n2),each=m2)
rho <- 0.2
offset1 <- offset2 <- runif(1,0,3)
sigmaMat1 <- ifelse(diag(m1)==1,1,rho)
sigmaMat2 <- ifelse(diag(m2)==1,1,rho)
t1 <- runif(m1,0,24)
tt1 <- rep(t1,n1)
yy1 <- as.vector(t(mvrnorm(n1,rep(offset1,m1),sigmaMat1)))
t2 <- runif(m2,0,24)
tt2 <- rep(t2,n2)
yy2 <- as.vector(t(mvrnorm(n2,rep(offset2,m2),sigmaMat2)))
group <- c(rep("group1",n1*m1),rep("group2",n2*m2))
rpt_diff(tt1,yy1,tt2,yy2,id1,id2,group)
}
\author{
Haocheng Ding, Zhiguang Huo
}
