% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_rhythmicity.R
\name{rpt_rhythmicity}
\alias{rpt_rhythmicity}
\title{Likelihood-based Test for Detecting Circadian Pattern with Repeated Measurement.}
\usage{
rpt_rhythmicity(tt, yy, id, period = 24, method = "LR")
}
\arguments{
\item{tt}{Time vector}

\item{yy}{Expression vector}

\item{id}{Subject ID number}

\item{period}{Period of the since curve. Default is 24.}

\item{method}{Testing methods can be "LR" or "F". Default is "LR".}
}
\value{
A list of test statistic and pvalue.
Formula 1: \eqn{yy = amp * sin(2\pi/period * (phase + tt)) + offset}
Formula 2: \eqn{yy = A * sin(2\pi/period * tt) + B * cos(2*\pi/period * tt) + offset}
\item{stat}{Test statistic.}
\item{pvalue}{P-value from the test.}
}
\description{
Likelihood-based test for circadian pattern detection with repeated measurement.
}
\details{
Test the significance of circadian curve fitting using mixed model with random intercept and likelihood-based test.
}
\examples{
Example 1

set.seed(32611)
m <- 10
n <- 12
id <- rep(1:n,each=m)
rho <- 0.2
offset <- runif(1,0,3)
sigmaMat <- ifelse(diag(m)==1,1,rho)
tt <- rep(runif(m,0,24),n)
yy <- as.vector(t(mvrnorm(n,rep(offset,m),sigmaMat)))
rpt_rhythmicity(tt, yy, id, period=24, method="LR")

Example 2
set.seed(32611)
m <- 10
n <- 12
id <- rep(1:n,each=m)
rho <- 0.2
amp <- 1
phase <- 3
offset <- runif(1,0,3)
sigmaMat <- ifelse(diag(m)==1,1,rho)
t <- runif(m,0,24)
tt <- rep(t,n)
yy <- as.vector(t(mvrnorm(n,amp*sin(2*pi/24*(phase+t))+offset,sigmaMat)))
rpt_rhythmicity(tt, yy, id, period=24, method="LR")
}
\author{
Haocheng Ding, Zhiguang Huo
}
