% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subHTRF.R
\docType{data}
\name{subHTRF}
\alias{subHTRF}
\title{Human time-restricted feeding subdata}
\format{
A data frame with 1195 rows and 128 variables. Each row is gene expression value. First three columns are three types of gene identification names then each column represents one sample.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE129843}
}
\usage{
subHTRF
}
\description{
Subset of Short-Term Time-Restricted Feeding Alters Rhythmicity of Lipid and Amino Acid Metabolites in Overweight Men Data (Human time-restricted feeding subdata)
}
\references{
Lundell LS, Parr EB, Devlin BL, Ingerslev LR et al. Time-restricted feeding alters lipid and amino acid metabolite rhythmicity without perturbing clock gene expression. Nat Commun 2020 Sep 16;11(1):4643. PMID: 32938935.
}
\keyword{datasets}
